refactor: move gallery and best of movies packages in cast details package
