+++ app/src/main/java/com/moscow/cineverse/MainActivity.kt
@@ -8,7 +8,7 @@ import com.google.firebase.analytics.FirebaseAnalytics
 import com.google.firebase.analytics.ktx.analytics
 import com.google.firebase.ktx.Firebase
 import com.moscow.cineverse.designSystem.theme.CineVerseTheme
-import com.moscow.cineverse.screen.cast_details_gallery.ActorGalleryScreen
+import com.moscow.cineverse.screen.castDetails.gallery.ActorGalleryScreen
 
 class MainActivity : ComponentActivity() {
