+++ presentation/src/main/java/com/moscow/cineverse/screen/castDetails/best0fmovies/ShowAllActorMoviesEvents.kt
@@ -1,0 +1,4 @@
+package com.moscow.cineverse.screen.castDetails.best0fmovies
+
+sealed class ShowAllActorMoviesEvents {
+}
