+++ presentation/src/main/java/com/moscow/cineverse/screen/castDetails/best0fmovies/ShowAllActorMoviesScreen.kt
@@ -1,13 +1,17 @@
-package com.moscow.cineverse.screen.cast_details_gallery
+package com.moscow.cineverse.screen.castDetails.best0fmovies
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
@@ -20,33 +24,35 @@ import com.moscow.cineverse.designSystem.component.MovieButton
 import com.moscow.cineverse.designSystem.component.MovieCircularProgressBar
 import com.moscow.cineverse.designSystem.component.MovieScaffold
 import com.moscow.cineverse.designSystem.component.MovieText
-import com.moscow.cineverse.designSystem.component.cast_details.CastGallery
+import com.moscow.cineverse.designSystem.component.ViewMode
+import com.moscow.cineverse.designSystem.component.ViewModeToggle
 import com.moscow.cineverse.designSystem.theme.Theme
+import com.moscow.cineverse.screen.component.movie_poster_card.MediaItemUi
+import com.moscow.cineverse.screen.component.movie_poster_card.MoviePosterCard
 import com.moscow.cinverse.presentation.R
 import org.koin.androidx.compose.koinViewModel
-import org.koin.core.parameter.parametersOf
 
 @Composable
-fun ActorGalleryScreen(
-    actorId: Int,
-    title: String,
+fun ShowAllActorMoviesScreen(
     modifier: Modifier = Modifier,
+    title: String
 ) {
-    val viewModel: ActorGalleryViewModel = koinViewModel(parameters = { parametersOf(actorId, title) })
+    val viewModel: ShowAllActorMoviesInteractionViewModel = koinViewModel()
     val uiState by viewModel.uiState.collectAsState()
 
-    ActorGalleryContent(
+    ShowAllActorMoviesContent(
         uiState = uiState,
         interactionListener = viewModel,
         modifier = modifier,
-        title = uiState.actorName
+        title = title
     )
 }
 
+
 @Composable
-fun ActorGalleryContent(
+fun ShowAllActorMoviesContent(
     uiState: ShowAllActorMoviesState,
-    interactionListener: ActorGalleryInteractionListener,
+    interactionListener: ShowAllActorMoviesInteractionListener,
     modifier: Modifier = Modifier,
     title: String
 ) {
@@ -92,35 +98,75 @@ fun ActorGalleryContent(
                             title = title,
                             backButtonClick = interactionListener::backButtonClick,
                         )
-                    CastGallery(
-                        images = uiState.photos,
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .padding(horizontal = 16.dp),
-                    )
-                }
+                        LazyVerticalGrid(
+                            columns = if (uiState.viewMode == ViewMode.GRID)
+                                GridCells.Fixed(2)
+                            else
+                                GridCells.Fixed(1),
+                            contentPadding = PaddingValues(
+                                vertical = 16.dp,
+                                horizontal = 16.dp
+                            ),
+                            verticalArrangement = Arrangement.spacedBy(16.dp),
+                            horizontalArrangement = Arrangement.spacedBy(12.dp),
+                            modifier = Modifier.fillMaxSize()
+                        ) {
+                            items(uiState.movies) { item ->
+                                MoviePosterCard(
+                                    movie = item,
+                                    viewMode = uiState.viewMode,
+                                    onMovieClick = interactionListener::onMovieClick
+                                )
+
+
+                            }
+                        }
                     }
+                    ViewModeToggle(
+                        selectedMode = uiState.viewMode,
+                        onModeSelected = interactionListener::onViewModeChanged,
+                        modifier = Modifier
+                            .align(Alignment.BottomEnd)
+                            .padding(16.dp)
+                    )
+
+                }
             }
 
 
         }
     }
+
+
 }
 
 @CineVersePreviews
 @Composable
-fun ActorGalleryPreview(modifier: Modifier = Modifier) {
-    ActorGalleryContent(
+fun ShowAllActorMoviesPreview(modifier: Modifier = Modifier) {
+    ShowAllActorMoviesContent(
         uiState = ShowAllActorMoviesState(
             isLoading = false,
             error = null,
-            photos = emptyList()
+            viewMode = ViewMode.GRID,
+            movies = List(20) { index ->
+                MediaItemUi(
+                    id = index,
+                    title = "Movie $index",
+                    posterPath = "https://example.com/poster_$index.jpg",
+                    rating = 7.5f,
+                    genres = listOf("Action", "Adventure"),
+                    duration = "2h 30m",
+                    releaseDate = "2023-10-01"
+                )
+            }
         ),
-        interactionListener = object : ActorGalleryInteractionListener {
+        interactionListener = object : ShowAllActorMoviesInteractionListener {
             override fun onRefresh() {}
+            override fun onViewModeChanged(viewMode: ViewMode) {}
+            override fun onMovieClick(movieId: Int) {}
             override fun backButtonClick() {}
         },
         modifier = modifier,
-        title = "Actor Gallery"
+        "sad"
     )
 }
