+++ presentation/src/main/java/com/moscow/cineverse/screen/castDetails/best0fmovies/ShowAllActorMoviesState.kt
@@ -1,4 +1,4 @@
-package com.moscow.cineverse.screen.cast_details_show_all
+package com.moscow.cineverse.screen.castDetails.best0fmovies
 
 import com.moscow.cineverse.designSystem.component.ViewMode
 import com.moscow.cineverse.screen.component.movie_poster_card.MediaItemUi
