+++ presentation/src/main/java/com/moscow/cineverse/screen/castDetails/gallery/ActorGalleryEvents.kt
@@ -1,0 +1,5 @@
+package com.moscow.cineverse.screen.castDetails.gallery
+
+sealed class ActorGalleryEvents {
+
+}
