+++ presentation/src/main/java/com/moscow/cineverse/screen/castDetails/gallery/ActorGalleryScreen.kt
@@ -1,16 +1,11 @@
-package com.moscow.cineverse.screen.cast_details_show_all
+package com.moscow.cineverse.screen.castDetails.gallery
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -24,35 +19,33 @@ import com.moscow.cineverse.designSystem.component.MovieButton
 import com.moscow.cineverse.designSystem.component.MovieCircularProgressBar
 import com.moscow.cineverse.designSystem.component.MovieScaffold
 import com.moscow.cineverse.designSystem.component.MovieText
-import com.moscow.cineverse.designSystem.component.ViewMode
-import com.moscow.cineverse.designSystem.component.ViewModeToggle
+import com.moscow.cineverse.designSystem.component.cast_details.CastGallery
 import com.moscow.cineverse.designSystem.theme.Theme
-import com.moscow.cineverse.screen.component.movie_poster_card.MediaItemUi
-import com.moscow.cineverse.screen.component.movie_poster_card.MoviePosterCard
 import com.moscow.cinverse.presentation.R
 import org.koin.androidx.compose.koinViewModel
+import org.koin.core.parameter.parametersOf
 
 @Composable
-fun ShowAllActorMoviesScreen(
+fun ActorGalleryScreen(
+    actorId: Int,
+    title: String,
     modifier: Modifier = Modifier,
-    title: String
 ) {
-    val viewModel: ShowAllActorMoviesInteractionViewModel = koinViewModel()
+    val viewModel: ActorGalleryViewModel = koinViewModel(parameters = { parametersOf(actorId, title) })
     val uiState by viewModel.uiState.collectAsState()
 
-    ShowAllActorMoviesContent(
+    ActorGalleryContent(
         uiState = uiState,
         interactionListener = viewModel,
         modifier = modifier,
-        title = title
+        title = uiState.actorName
     )
 }
 
-
 @Composable
-fun ShowAllActorMoviesContent(
+fun ActorGalleryContent(
     uiState: ShowAllActorMoviesState,
-    interactionListener: ShowAllActorMoviesInteractionListener,
+    interactionListener: ActorGalleryInteractionListener,
     modifier: Modifier = Modifier,
     title: String
 ) {
@@ -98,75 +91,35 @@ fun ShowAllActorMoviesContent(
                             title = title,
                             backButtonClick = interactionListener::backButtonClick,
                         )
-                        LazyVerticalGrid(
-                            columns = if (uiState.viewMode == ViewMode.GRID)
-                                GridCells.Fixed(2)
-                            else
-                                GridCells.Fixed(1),
-                            contentPadding = PaddingValues(
-                                vertical = 16.dp,
-                                horizontal = 16.dp
-                            ),
-                            verticalArrangement = Arrangement.spacedBy(16.dp),
-                            horizontalArrangement = Arrangement.spacedBy(12.dp),
-                            modifier = Modifier.fillMaxSize()
-                        ) {
-                            items(uiState.movies) { item ->
-                                MoviePosterCard(
-                                    movie = item,
-                                    viewMode = uiState.viewMode,
-                                    onMovieClick = interactionListener::onMovieClick
-                                )
-
-
-                            }
-                        }
-                    }
-                    ViewModeToggle(
-                        selectedMode = uiState.viewMode,
-                        onModeSelected = interactionListener::onViewModeChanged,
+                    CastGallery(
+                        images = uiState.photos,
                         modifier = Modifier
-                            .align(Alignment.BottomEnd)
-                            .padding(16.dp)
+                            .fillMaxSize()
+                            .padding(horizontal = 16.dp),
                     )
-
                 }
+                    }
             }
 
 
         }
     }
-
-
 }
 
 @CineVersePreviews
 @Composable
-fun ShowAllActorMoviesPreview(modifier: Modifier = Modifier) {
-    ShowAllActorMoviesContent(
+fun ActorGalleryPreview(modifier: Modifier = Modifier) {
+    ActorGalleryContent(
         uiState = ShowAllActorMoviesState(
             isLoading = false,
             error = null,
-            viewMode = ViewMode.GRID,
-            movies = List(20) { index ->
-                MediaItemUi(
-                    id = index,
-                    title = "Movie $index",
-                    posterPath = "https://example.com/poster_$index.jpg",
-                    rating = 7.5f,
-                    genres = listOf("Action", "Adventure"),
-                    duration = "2h 30m",
-                    releaseDate = "2023-10-01"
-                )
-            }
+            photos = emptyList()
         ),
-        interactionListener = object : ShowAllActorMoviesInteractionListener {
+        interactionListener = object : ActorGalleryInteractionListener {
             override fun onRefresh() {}
-            override fun onViewModeChanged(viewMode: ViewMode) {}
-            override fun onMovieClick(movieId: Int) {}
             override fun backButtonClick() {}
         },
         modifier = modifier,
-        "sad"
+        title = "Actor Gallery"
     )
 }
