+++ presentation/src/main/java/com/moscow/cineverse/screen/castDetails/gallery/ActorGalleryState.kt
@@ -1,4 +1,4 @@
-package com.moscow.cineverse.screen.cast_details_gallery
+package com.moscow.cineverse.screen.castDetails.gallery
 
 data class ShowAllActorMoviesState(
     val actorId: Int = 0,
