--- presentation/src/main/java/com/moscow/cineverse/screen/cast_details_gallery/ActorGalleryEvents.kt
@@ -1,5 +1,0 @@
-package com.moscow.cineverse.screen.cast_details_gallery
-
-sealed class ActorGalleryEvents {
-
-}
