--- presentation/src/main/java/com/moscow/cineverse/screen/cast_details_show_all/ShowAllActorMoviesEvents.kt
@@ -1,4 +1,0 @@
-package com.moscow.cineverse.screen.cast_details_show_all
-
-sealed class ShowAllActorMoviesEvents {
-}
